Add similar images from map to local SfM
This approach select similar images to local SfM images in global map.
Filter images with small number of correspondences and try to register
found images into the local SfM.
This approch does not work because of low accuracy of the localization.
TODO: add the generalized absolute pose solver and localize the whole
local SfM w.r.t. the global map.